# Conflicts:
#	bootstrap/bungeecord/src/main/java/org/geysermc/geyser/platform/bungeecord/GeyserBungeePlugin.java
#	bootstrap/bungeecord/src/main/java/org/geysermc/geyser/platform/bungeecord/command/BungeeCommandSource.java
#	bootstrap/bungeecord/src/main/java/org/geysermc/geyser/platform/bungeecord/command/GeyserBungeeCommandExecutor.java
#	bootstrap/spigot/src/main/java/org/geysermc/geyser/platform/spigot/GeyserSpigotPlugin.java
#	bootstrap/spigot/src/main/java/org/geysermc/geyser/platform/spigot/command/GeyserSpigotCommandExecutor.java
#	bootstrap/spigot/src/main/java/org/geysermc/geyser/platform/spigot/command/SpigotCommandSource.java
#	bootstrap/sponge/pom.xml
#	bootstrap/sponge/src/main/java/org/geysermc/geyser/platform/sponge/GeyserSpongePingPassthrough.java
#	bootstrap/sponge/src/main/java/org/geysermc/geyser/platform/sponge/GeyserSpongePlugin.java
#	bootstrap/sponge/src/main/java/org/geysermc/geyser/platform/sponge/command/GeyserSpongeCommandExecutor.java
#	bootstrap/sponge/src/main/java/org/geysermc/geyser/platform/sponge/command/GeyserSpongeCommandManager.java
#	bootstrap/sponge/src/main/java/org/geysermc/geyser/platform/sponge/command/SpongeCommandSource.java
#	bootstrap/standalone/src/main/java/org/geysermc/geyser/platform/standalone/GeyserStandaloneBootstrap.java
#	bootstrap/standalone/src/main/java/org/geysermc/geyser/platform/standalone/GeyserStandaloneLogger.java
#	bootstrap/standalone/src/main/java/org/geysermc/geyser/platform/standalone/gui/GeyserStandaloneGUI.java
#	bootstrap/velocity/pom.xml
#	bootstrap/velocity/src/main/java/org/geysermc/geyser/platform/velocity/GeyserVelocityPlugin.java
#	bootstrap/velocity/src/main/java/org/geysermc/geyser/platform/velocity/command/GeyserVelocityCommandExecutor.java
#	core/src/main/java/org/geysermc/geyser/GeyserImpl.java
#	core/src/main/java/org/geysermc/geyser/command/GeyserCommandExecutor.java
#	core/src/main/java/org/geysermc/geyser/command/GeyserCommandManager.java
#	core/src/main/java/org/geysermc/geyser/command/GeyserCommandSource.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/AdvancedTooltipsCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/AdvancementsCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/ConnectionTestCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/DumpCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/HelpCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/ListCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/OffhandCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/ReloadCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/SettingsCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/StatisticsCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/StopCommand.java
#	core/src/main/java/org/geysermc/geyser/command/defaults/VersionCommand.java
#	core/src/main/java/org/geysermc/geyser/session/GeyserSession.java
#	core/src/main/java/org/geysermc/geyser/translator/protocol/bedrock/BedrockCommandRequestTranslator.java
#	core/src/main/java/org/geysermc/geyser/translator/protocol/java/JavaCommandsTranslator.java

+++ core/src/main/java/org/geysermc/geyser/registry/populator/BlockRegistryPopulator.java
@@ -234,7 +234,7 @@ public final class BlockRegistryPopulator {
             BlockMapping.BlockMappingBuilder builder = BlockMapping.builder();
             JsonNode hardnessNode = entry.getValue().get("block_hardness");
             if (hardnessNode != null) {
-                builder.hardness(hardnessNode.doubleValue());
+                builder.hardness(hardnessNode.floatValue());
             }
 
             JsonNode canBreakWithHandNode = entry.getValue().get("can_break_with_hand");
